authorize successfully

+++ src/main/java/fu/swp/dorm_mnm/repository/UserRepository.java
@@ -1,13 +1,14 @@
 package fu.swp.dorm_mnm.repository;
 
 import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import fu.swp.dorm_mnm.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long>{
-    Optional<User> findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+  Optional<User> findByUsername(String username);
+
 }
